cliently_backend/requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
load-dotenv==0.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
